+++ app/src/main/java/ua/uz/vopak/brb4/brb4/models/DocWaresModelIncome.java
@@ -3,6 +3,6 @@ package ua.uz.vopak.brb4.brb4.models;
 public class DocWaresModelIncome {
     public String CodeWares;
     public String NameWares;
-    public Integer QuantityOrdered;
-    public Integer QuantityIncoming;
+    public Float QuantityOrdered;
+    public Float QuantityIncoming;
 }
